+++ source/DiskImageHelper.h
@@ -393,7 +393,7 @@ protected:
 	void GetCharLowerExt2(char* pszExt, LPCTSTR pszImageFilename, const UINT uExtSize);
 	void SetImageInfo(ImageInfo* pImageInfo, FileType_e fileType, uint32_t dwOffset, CImageBase* pImageType, uint32_t dwSize);
 
-	UINT GetNumImages(void) { return m_vecImageTypes.size(); };
+	UINT GetNumImages(void) { return (UINT) m_vecImageTypes.size(); };
 	CImageBase* GetImage(UINT uIndex) { _ASSERT(uIndex<GetNumImages()); return m_vecImageTypes[uIndex]; }
 	CImageBase* GetImage(eImageType Type)
 	{
